+++ packages/atlas/src/views/viewer/ChannelView/ChannelViewTabs/index.ts
@@ -1,0 +1,3 @@
+export * from './ChannelAbout'
+export * from './ChannelNfts'
+export * from './ChannelVideos'
